First commit contains tasks usurped from Marcin Grzejszczak's
Spring Cloud Pipelines.

Spring Cloud Pipelines is very opinionated in some of its
assumptions. For example, it assumes every application is
a spring-boot microservice that uses spring-cloud and
spring-cloud-contract to verify the contracts between them.
We have found that typically it requires heavy customization
of Spring Cloud Pipelines before it is usable at a customer
location.

We are therefore going to make the assumption that most
engagements are going to need to make their own pipelines
that contain some opinions that do not align with those of
SCP. So this repo will provide tasks that will allow you to
create your own pipelines with well-defined extension points
in places where the task functionality can be customized.